added busines logic into CoffeeController and change max Name lenght in CoffeeShopContex

+++ CoffeeShop/Controllers/CoffeeController.cs
@@ -26,13 +26,33 @@ namespace CoffeeShop.Controllers
         [HttpPost]
         public void PostNewCoffee([FromBody]Coffee coffee)
         {
+            if (db.Coffees.Any(x => x.Id == coffee.Id))
+            {
+                throw new ArgumentException("A coffee with this id already exists", nameof(coffee));
+            }
+
+            if (!IsCoffeeValid(coffee))
+            {
+                throw new ArgumentException("Entered data is not correct", nameof(coffee));
+            }
+
             db.Coffees.Add(coffee);
             db.SaveChanges();
         }
 
         [HttpPut]
-        public void PutById(int id, [FromBody] Coffee coffee)
+        public void PutById(int id, [FromBody]Coffee coffee)
         {
+            if (!db.Coffees.Any(x => x.Id == coffee.Id))
+            {
+                throw new ArgumentException("Coffee with this id does not exist", nameof(coffee));
+            }
+
+            if (!IsCoffeeValid(coffee))
+            {
+                throw new ArgumentException("Entered data is not correct", nameof(coffee));
+            }
+
             if (id == coffee.Id)
             {
                 db.Entry(coffee).State = EntityState.Modified;
@@ -43,12 +63,33 @@ namespace CoffeeShop.Controllers
         [HttpDelete]
         public void DeleteById(int id)
         {
+            if (!db.Coffees.Any(x => x.Id == id))
+            {
+                throw new ArgumentException("Coffee with this id does not exist", nameof(id));
+            }
+
             Coffee coffee = db.Coffees.Find(id);
+
             if (coffee != null)
             {
                 db.Coffees.Remove(coffee);
                 db.SaveChanges();
             }
+        }
+
+        private bool IsCoffeeValid(Coffee coffee)
+        {
+            if (string.IsNullOrWhiteSpace(coffee.Name) || coffee.Name.Length > 30)
+            {
+                return false;
+            }
+
+            if (coffee.Price <= 0)
+            {
+                return false;
+            }
+
+            return true;
         }
     }
 }
